+++ packages/frontend/components/panes/EmptyMain.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import InformationCircleIcon from '@heroicons/react/solid/InformationCircleIcon';
+import { InformationCircleIcon } from '@heroicons/react/24/solid';
 import { Icon, VStack, Text } from '@chakra-ui/react';
 
 export const EmptyMain: React.FC = (props) => {