+++ packages/frontend/pages/settings/components/customize/CustomizeAppearance.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
 import { Box, Center, Icon, Radio, Stack } from '@chakra-ui/react';
 import { Formik, Form } from 'formik';
-import MoonIcon from '@heroicons/react/solid/MoonIcon';
-import SunIcon from '@heroicons/react/solid/SunIcon';
+import { MoonIcon, SunIcon } from '@heroicons/react/24/solid';
 import {
   AvatarStyle,
-  AvatarStyles,
   CollectionLayout,
-  CollectionLayouts,
   defaultArticleFontFamily,
   defaultArticleFontSize,
   defaultArticleMonoFontFamily,
@@ -92,7 +89,7 @@ export const CustomizeAppearance: React.FC<CustomizeAppearanceProps> = (
                 name="defaultCollectionLayout"
               >
                 {({ getRadioProps }) =>
-                  CollectionLayouts.map((value: CollectionLayout) => {
+                  CollectionLayout.options.map((value: CollectionLayout) => {
                     const radio = getRadioProps({ value });
                     return (
                       <Radio key={value} {...radio}>
@@ -105,7 +102,7 @@ export const CustomizeAppearance: React.FC<CustomizeAppearanceProps> = (
 
               <RadioGroupField label="Avatar style" name="avatarStyle">
                 {({ getRadioProps }) =>
-                  AvatarStyles.map((value: AvatarStyle) => {
+                  AvatarStyle.options.map((value: AvatarStyle) => {
                     const radio = getRadioProps({ value });
                     return (
                       <Radio key={value} {...radio}>